requirements.txt
mkdocs-with-pdf
mkdocs-simple-hooks
pyyaml
